+++ repos/os/src/app/ping/xml_node.cc
@@ -17,11 +17,11 @@
 using namespace Genode;
 
 
-Microseconds Genode::read_sec_attr(Xml_node      const  node,
-                                   char          const *name,
-                                   unsigned long const  default_sec)
+Microseconds Genode::read_sec_attr(Xml_node const  node,
+                                   char     const *name,
+                                   uint64_t const  default_sec)
 {
-	unsigned long sec = node.attribute_value(name, 0UL);
+	uint64_t sec = node.attribute_value(name, (uint64_t)0);
 	if (!sec) {
 		sec = default_sec;
 	}
