+++ repos/os/src/drivers/gpio/spec/rpi/gpio.h
@@ -117,7 +117,7 @@ class Gpio::Reg : Attached_io_mem_dataspace, Mmio
 			/**
 			 * Implementation of 'Delayer' interface
 			 */
-			void usleep(unsigned us) override { Timer::Connection::usleep(us); }
+			void usleep(uint64_t us) override { Timer::Connection::usleep(us); }
 
 		} _delayer;
