+++ repos/os/src/test/init/main.cc
@@ -279,7 +279,7 @@ struct Test::Main : Log_message_handler
 
 			if (step.type() == "sleep") {
 				if (!_timer_scheduled) {
-					unsigned long const timeout_ms = step.attribute_value("ms", 250UL);
+					uint64_t const timeout_ms = step.attribute_value("ms", (uint64_t)250);
 					_timer.trigger_once(timeout_ms*1000);
 					_timer_scheduled = true;
 				}
